ログインユーザーの場合のみ表示する処理、投稿ユーザー名の表示、ログアウト時のルーティングの見直しなど

+++ resources/views/answers/all_answers.blade.php
@@ -61,19 +61,24 @@
 			@foreach ($answers as $answer)
 				<div class="answers container text-center border border-dark border-2 rounded-3 mb-3">
 					<div class='answer'>
+						<p class="row align-items-start">回答ユーザー：{{ $answer->user_name }}</p>
 						<div class="d-flex justify-content-between">
 							<p class="row align-items-start">{{ $answer->created_at }}</p>
 						</div>
-						<div class="d-flex justify-content-between">
-							<a href="/answers/{{ $answer->id }}/edit_answer">回答の編集</a>
-						</div>
-						<div class="d-flex justify-content-between">
-							<form action="/answers/{{ $answer->id }}" id="form_{{ $answer->id }}" method="post">
-								@csrf
-								@method('DELETE')
-								<a href="#" onclick="deleteAnswer({{ $answer->id }})" style="color:red">回答の削除</a>
-							</form>
-						</div>
+						@foreach($answer->users as $user)
+							@if (Auth::user()->id == $user->id)
+								<div class="d-flex justify-content-between">
+										<a href="/answers/{{ $answer->id }}/edit_answer">回答の編集</a>
+								</div>
+								<div class="d-flex justify-content-between">
+									<form action="/answers/{{ $answer->id }}" id="form_{{ $answer->id }}" method="post">
+										@csrf
+										@method('DELETE')
+										<a href="#" onclick="deleteAnswer({{ $answer->id }})" style="color:red">回答の削除</a>
+									</form>
+								</div>
+				  			@endif
+			  			@endforeach
 						<p class='body row align-items-start'>{{ $answer->body }}</p>
 						@if(strrpos($answer->file_path, '.png'))
 						    <img src="{{ $answer->file_path }}">
